fix(Healthcheck): Fixues docker healthcheck
Points dockerfile to location of new healthcheck script

Fixes #699
